changes are updated

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,13 +10,13 @@
 size_t print_list(const list_t *h)
 {
 	const list_t *nodo;
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	nodo = h;
 	while (nodo != NULL)
 	{
-	printf("[%u] %s\n", nodo->len, nodo->str);
+	printf("[%d] %s\n", nodo->len, nodo->str);
 	nodo = nodo->next;
 	i++;
 	}
